code refactoring as per last pr comment
