+++ src/controllers/userController.ts
@@ -1,6 +1,7 @@
 import { Request, Response } from 'express';
 import mongoose from 'mongoose'; 
 import { isNil, isEmpty, some } from 'lodash';
+
 import * as userService from '@services/userService';
 import { HTTP_STATUS } from '@utils/constants';
 import { MESSAGES } from '@utils/message';
