+++ src/middleware/authMiddleware.ts
@@ -1,5 +1,7 @@
 import { Request, Response, NextFunction } from 'express';
 import jwt from 'jsonwebtoken';
+import { isEmpty } from 'lodash';
+
 import { HTTP_STATUS, SECRET_KEY } from '@utils/constants';
 import { MESSAGES } from '@utils/message';
 import { sendResponse } from '@utils/Respons/response';
@@ -11,6 +13,9 @@ export const authenticateUser = (req: AuthRequest, res: Response, next: NextFunc
     return sendResponse(res, HTTP_STATUS.UNAUTHORIZED, MESSAGES.UNAUTHORIZED);
   }  
   const token = authHeader.split(' ')[1];
+  if (isEmpty(token)) {
+    return sendResponse(res, HTTP_STATUS.UNAUTHORIZED, MESSAGES.UNAUTHORIZED);
+  }
   try {
     const decoded = jwt.verify(token, `${SECRET_KEY}`) as { userId: string };
     req.userId = decoded.userId;
