+++ src/routes/resultRoutes.ts
@@ -1,6 +1,7 @@
 import express from 'express';
+
 import * as resultController from '@controllers/resultController';
-import { authenticateUser } from 'middleware/authMiddleware';
+import { authenticateUser } from '@middleware/authMiddleware';
 
 const resultRoutes = express.Router();
 
@@ -9,7 +10,6 @@ resultRoutes.get('/', authenticateUser, resultController.getResults);
 resultRoutes.get('/:resultId', authenticateUser, resultController.getResultById);
 resultRoutes.put('/:resultId', authenticateUser, resultController.updateResult);
 resultRoutes.delete('/:resultId', authenticateUser, resultController.deleteResult);    
-resultRoutes.get('/stats/overall', authenticateUser, resultController.getResultsStats);
 
 export default resultRoutes;
